refactor: change folder name
ONE-vscode-DCO-1.0-Signed-off-by: Yeseo Kang <yeeybook@gmail.com>
