+++ src/Config/ConfigurationSettingsPanel.ts
